Ok one request
Can you make clicking the menu button in the top right corner assign #citylist to have opacity 1 and z index 300?
I just want it to overlay when you hit the menu and then there will be a slize button somewhere to make it go away if you dont click anything.